Use a provider to construct ORSO dataset
