+++ src/essreflectometry/__init__.py
@@ -4,7 +4,7 @@
 # flake8: noqa: F401
 import importlib.metadata
 
-from . import calibrations, conversions, corrections, io, normalize
+from . import calibrations, conversions, corrections, normalize, reductions
 
 try:
     __version__ = importlib.metadata.version(__package__ or __name__)
@@ -16,6 +16,7 @@ providers = (
     *corrections.providers,
     *calibrations.providers,
     *normalize.providers,
+    *reductions.providers,
 )
 """
 List of providers for setting up a Sciline pipeline.
